weeklyExamples/pyproject.toml
[project]
name = "weeklyexamples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lab>=8.8",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.6.0",
    "notebook>=7.5.3",
    "pandas>=3.0.1",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.49",
]

[tool.uv]
link-mode = "copy"
